added blank derby database
added a blank derby database called ScedulingAppDB to the project folder. Everyone must download the database and save it to their derby folder on their computer if they want to edit it. When finished editing the datebase. The new version should be put back into the project folder and commited with pull request so that everyone can have the latest updated version of the database